Fix circular dependencies in playwright test files
