+++ openmetadata-ui/src/main/resources/ui/playwright/utils/user.ts
@@ -12,27 +12,20 @@
  */
 
 import { expect, Page } from '@playwright/test';
-import { adjectives, nouns } from '../constant/user';
-import { toastNotification, uuid } from './common';
+import { UserClass } from '../support/user/UserClass';
+import { getAuthContext, getToken, toastNotification } from './common';
 
-export const getRandomFirstName = () => {
-  return `${
-    adjectives[Math.floor(Math.random() * adjectives.length)]
-  }${uuid()}`;
-};
-export const getRandomLastName = () => {
-  return `${nouns[Math.floor(Math.random() * nouns.length)]}${uuid()}`;
-};
-export const generateRandomUsername = () => {
-  const firstName = getRandomFirstName();
-  const lastName = getRandomLastName();
-
-  return {
-    firstName,
-    lastName,
-    email: `${firstName}.${lastName}@example.com`,
-    password: 'User@OMD123',
+export const performUserLogin = async (browser, user: UserClass) => {
+  const page = await browser.newPage();
+  await user.login(page);
+  const token = await getToken(page);
+  const apiContext = await getAuthContext(token);
+  const afterAction = async () => {
+    await apiContext.dispose();
+    await page.close();
   };
+
+  return { page, apiContext, afterAction };
 };
 
 export const nonDeletedUserChecks = async (page: Page) => {
